Confirm deletion of the project

+++ project-manager-react/src/actions/ProjectActions.js
@@ -39,9 +39,16 @@ export const getProject = (id, history) => async (dispatch) => {
 };
 
 export const deleteProject = (id) => async (dispatch) => {
-  await axios.delete(`/api/project/${id}`);
-  dispatch({
-    type: DELETE_PROJECT,
-    payload: id,
-  });
+  if (
+    window.confirm(
+      "Are sure that you want delete project? \
+      This will delete project and all data related to it."
+    )
+  ) {
+    await axios.delete(`/api/project/${id}`);
+    dispatch({
+      type: DELETE_PROJECT,
+      payload: id,
+    });
+  }
 };
